requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
Django==5.0.6
django-ckeditor-5==0.2.13
django-crispy-forms==2.2
django-js-asset==2.2.0
django-tinymce==4.1.0
djangorestframework==3.15.2
gunicorn==22.0.0
jsmin==3.0.1
packaging==24.1
pillow==10.3.0
pytz==2024.1
sqlparse==0.5.0
tzdata==2024.1
